Split message sink implementation into focused modules
